deleted InterviewInterviewerController and moved methods to IneterviewController
