+++ src/test/java/com/revature/service/InterviewServiceTests.java
@@ -84,9 +84,7 @@ public class InterviewServiceTests {
 		 */
 	  
 	  }
-	 
-	
-	
+
 	@Test
 	public void testGetAllInterview() {
 		List<Interview> interviews = new ArrayList<>();
